+++ src/extension.ts
@@ -23,8 +23,7 @@ import * as vscode from "vscode";
  * That is, the registered `Activation Event` has happened. The
  * `Activation Events` are configured in `package.json`, in the
  * `activationEvents` field.
- *
- * @param _context The `vscode.ExtensionContext` to use.
+ * @param context The `vscode.ExtensionContext` to use.
  */
 export async function activate(context: vscode.ExtensionContext) {
     const outChannel = vscode.window.createOutputChannel(c.outputChannelName);