+++ test/osInteraction-test.ts
@@ -11,7 +11,6 @@
  * Tests for the `odInteraction` module.
  */
 
-import * as c from "../src/constants";
 import * as chai from "chai";
 import * as h from "../src/extension_helpers";
 import * as io from "../src/osInteraction";
@@ -147,17 +146,6 @@ mocha.describe("I/O Functions", () => {
             chai.assert.strictEqual(out.stdout, "LICENSE\n", "ls of LICENSE");
             chai.assert.strictEqual(out.stderr, "", "No output at stderr!");
             chai.assert.isUndefined(out.error, "No error calling ls!");
-        });
-        mocha.it("Output of `dune --version`", async () => {
-            // eslint-disable-next-line array-bracket-newline
-            const out = await io.runCommand(root, c.duneCmd, [
-                c.duneVersionArg,
-                // eslint-disable-next-line array-bracket-newline
-            ]);
-            chai.assert.isString(out.stdout, "Something like `3.6.2`");
-            chai.assert.isNotEmpty(out.stdout, "Something like `3.6.2`");
-            chai.assert.strictEqual(out.stderr, "", "No output at stderr!");
-            chai.assert.isUndefined(out.error, "No error calling dune!");
         });
     });
 });